Compare grid cell counts

Compare grid cell counts

Usage

comp_trks(
  sim_trks,
  stations,
  land_barrier,
  vis_graph,
  multi.grid,
  HSgrid,
  snap_tolerance
)

Arguments

sim_trks

Tracks to which modeled tracks should be compared

stations

A sf object comprised of receiver station locations with a buffer around them that represents the range of the receiver (polygons)

land_barrier

A sf polygon of a barrier object around which tracks should be re-routed

vis_graph

A visibility graph created from the barrier object @seealso pathroutr::prt_visgraph

multi.grid

boolean

HSgrid

When multi.grid = FALSE, a sf polygon grid; When multi.grid = TRUE, a list of sf polygon grids

snap_tolerance

The tolerance (in meters) at which an intersection between a station and a track should snap to the station centroid. It is recommended that the snap_tolerance be equal to the station buffer size.

Value

A data frame with counts and differences by grid cell ID ("gid")