Distance to shore and density of receivers

This function determines the distance to shore and the density of receivers for each grid cell.

Usage

d_shore_rcvs(km, study_site, land_barrier, epsg, sts_pts)

Arguments

km

grid cell resolution in km. one-sided length of grid cell, assumes desired grid cell is to be squared

study_site

simple feature polygon object that encompasses the entire study site

land_barrier

simple feature (multi)polygon object to route tracks around

epsg

epsg code for desired coordinated system transformation

sts_pts

a simple feature (multi)point object representing receiver locations

Value

A simple feature multipolygon object with information on distance to shore from grid cell center, receiver presence/absence, counts, and densities