Get odds ratios

This function calculates the odds ratios for each variable in your model.

Usage

get_or(model)

Arguments

model

A model object created from the nlme package

Value

A data frame object with a point estimate and lower and upper 95% confidence interval values associated with each covariate

Examples

# Get the odds ratio for each variable in the final model

load(system.file("extdata", "fit2.100km.Rda", package = "sporeg"))

odds_100km <- get_or(fit2.100km) %>%
dplyr::mutate(res_name = "100km")