Gaps characteristics

Gaps characteristics

Usage

gps_chr(df)

Arguments

df

data frame object consisting of results from iterative reconstruction process

Value

a data frame object with summary statistics that provide insight into the locations of the gaps in the network receiver array that were closed by the reconstruction process

Examples

# Apply gps_chr to list of grid resolutions

library(sporeg)
library(dplyr)
library(sf)
library(data.table)
load(system.file("extdata", "res.Rda", package = "sporeg"))

clsd_gp_chars <- lapply(res, gps_chr) %>%
data.table::rbindlist(., idcol = 'resolution') %>%
  left_join(tibble(resolution = 1:4,
                   res_name = c("100km", "50km", "25km", "10km")),
            by = "resolution") %>%
            dplyr::mutate(res_name = ordered(res_name, levels = c("100km", "50km", "25km", "10km")))